moved typing to Constants
