+++ types/Constants.d.ts
@@ -46,3 +46,15 @@ export class ShoukakuTrack {
         uri?: string;
     };
 }
+
+export interface DecodedTrack {
+  identifier: string;
+  isSeekable: boolean;
+  author: string;
+  length: number;
+  isStream: boolean;
+  position: number;
+  title: string;
+  uri: string;
+  sourceName: string;
+}
